fixed bug in jQuery (ajax called twice)
Is there no in-DOM solution for establishing a target element to load the response? For example:

<%= link_to "Show panda", "images/panda.gif", {remote: true, data-target: '#panda-holder' %>

Then every remote link could call something generic like:

$('[data-remote=true]').on('click', function (e) {
e.preventdefault();
$($(this).attr('data-remote')).load($(this).attr('href'));
});

Or it could be more elegantly worked into rails.js. This avoids splitting ajax behaviour into dozens of .js.erb files, and writing format.js for every method in the controller. Sure wish something like that was built into rails.js.

Turbolinks is a great solution and I'm using it in a re-write, but for now I have to deal with a rails 3.1.10 app with loads of .on and e.preventdefaults in application.js